Reorganize UI components
