+++ src/components/UI/WorldMap/WorldMap.tsx
@@ -3,18 +3,18 @@ import {
   clickWorldCell,
   loadWorldRegionIndex,
   toggleWorldMap,
-} from "../../redux/gameActions";
+} from "../../../redux/gameActions";
 import "./WorldMap.css";
-import { generateWorldMap } from "../../generate/world";
+import { generateWorldMap } from "../../../generate/world";
 import {
   visitAnyWorldRegion,
   worldMapCellsX,
   worldMapCellsY,
-} from "../../constants";
+} from "../../../constants";
 import { WorldMapCell } from "./WorldMapCell";
-import { GameState } from "../../types/gameStateTypes";
-import { areCellsAdjacent, getNumberArray } from "../../utility";
-import { saveGame } from "../../data/saveLoad";
+import { GameState } from "../../../types/gameStateTypes";
+import { areCellsAdjacent, getNumberArray } from "../../../utility";
+import { saveGame } from "../../../data/saveLoad";
 
 export function WorldMap() {
   const style = {
